wip: optimize stack

+++ crates/tinywasm/src/runtime/stack/block_stack.rs
@@ -18,15 +18,15 @@ impl BlockStack {
 
     #[inline]
     /// get the label at the given index, where 0 is the top of the stack
-    pub(crate) fn get_relative_to(&self, index: usize, offset: usize) -> Option<&BlockFrame> {
-        let len = self.0.len() - offset;
+    pub(crate) fn get_relative_to(&self, index: u32, offset: u32) -> Option<&BlockFrame> {
+        let len = (self.0.len() as u32) - offset;
 
         // the vast majority of wasm functions don't use break to return
         if unlikely(index >= len) {
             return None;
         }
 
-        Some(&self.0[self.0.len() - index - 1])
+        Some(&self.0[self.0.len() - index as usize - 1])
     }
 
     #[inline]
@@ -36,31 +36,32 @@ impl BlockStack {
 
     /// keep the top `len` blocks and discard the rest
     #[inline]
-    pub(crate) fn truncate(&mut self, len: usize) {
-        self.0.truncate(len);
+    pub(crate) fn truncate(&mut self, len: u32) {
+        self.0.truncate(len as usize);
     }
 }
 
 #[derive(Debug, Clone)]
 pub(crate) struct BlockFrame {
     // position of the instruction pointer when the block was entered
-    pub(crate) instr_ptr: usize,
+    pub(crate) instr_ptr: u32,
     // position of the end instruction of the block
-    pub(crate) end_instr_ptr: usize,
+    pub(crate) end_instr_ptr: u32,
 
     // position of the stack pointer when the block was entered
-    pub(crate) stack_ptr: usize,
-    pub(crate) results: usize,
-    pub(crate) params: usize,
+    pub(crate) stack_ptr: u32,
+
+    pub(crate) results: u8,
+    pub(crate) params: u8,
     pub(crate) ty: BlockType,
 }
 
 impl BlockFrame {
     #[inline]
     pub(crate) fn new(
-        instr_ptr: usize,
-        end_instr_ptr: usize,
-        stack_ptr: usize,
+        instr_ptr: u32,
+        end_instr_ptr: u32,
+        stack_ptr: u32,
         ty: BlockType,
         args: &BlockArgs,
         module: &ModuleInstance,
@@ -70,7 +71,7 @@ impl BlockFrame {
             BlockArgs::Type(_) => (0, 1),
             BlockArgs::FuncType(t) => {
                 let ty = module.func_ty(*t);
-                (ty.params.len(), ty.results.len())
+                (ty.params.len() as u8, ty.results.len() as u8)
             }
         };
